Make title of page editable
This change puts the `<title>` element in it's own block so it can be changed by users.  

Something as subjective as the page title should not be hardcoded into the template IMO.

I created a new block instead of reusing the `meta` block to avoid breaking the title on templates created prior to this change.